New sparse_nlmap DEM method
This method allows computing DEMs from multiple instruments with differing spatial resolution without compromising the resolution of either one. 'Sparse' in this case refers to use of sparse matrix methods to enable this functionality, rather than 'sparsity' of the solution vector as a constraint as in Mark Cheung's sparse_em basis pursuit method. The mathematical terminology in this case is degenerate and there isn't a good alternative word to refer to sparse matrices.